ina219: Interpret signed output register values (#54)
The current and shunt voltage register values are signed, twos
complement format. In my tests interpreting them directly in the
way implemented here gave output consistent with other ina219
libraries.

+++ ina219/ina219.go
@@ -110,7 +110,7 @@ func (d *Dev) Sense() (PowerMonitor, error) {
 		return PowerMonitor{}, errReadShunt
 	}
 	// Least significant bit is 10µV.
-	pm.Shunt = physic.ElectricPotential(shunt) * 10 * physic.MicroVolt
+	pm.Shunt = physic.ElectricPotential(int16(shunt)) * 10 * physic.MicroVolt
 
 	bus, err := d.m.ReadUint16(busVoltageRegister)
 	if err != nil {
@@ -128,7 +128,7 @@ func (d *Dev) Sense() (PowerMonitor, error) {
 	if err != nil {
 		return PowerMonitor{}, errReadCurrent
 	}
-	pm.Current = physic.ElectricCurrent(current) * d.currentLSB
+	pm.Current = physic.ElectricCurrent(int16(current)) * d.currentLSB
 
 	power, err := d.m.ReadUint16(powerRegister)
 	if err != nil {
